Block root usage

+++ forever.sh
@@ -1,7 +1,8 @@
 #!/usr/bin/env bash
 
 set -euo pipefail
-set -x
+
+if [ "$(id -u)" -eq 0 ]; then echo "Please don't run as root." >&2; exit 1; fi
 
 rm -vf src/com/github/ivarref/*.class 2> /dev/null || true
